Fix unfound template when service failed in update action
On service failure, it was rendering action: :edit which expect to find the template at the default location for the edit action: modules/storages/app/views/storages/admin/storages/edit.html.erb. As we put it in a non standard place (modules/storages/app/views/storages/admin/edit.html.erb) it was not found and produced an error.

Sticking to default locations fixed the issue and allowed removing some useless code.