Fix team planner crash when adding a wp from the add existing pane
When adding from the add existing pane, the assignee will get set to an object that only has the href set without
a name or avatar attribute. When the patch request to update the wp fails, this crashes the principal renderer since
there's only half an assignee available (really this change should be unpatched)

Here is the easy fix; we just don't try rendering an avatar if there's no name. This makes sure the team planner stays
in a usable state.