Use Cases
I added a new folder/repository called "Use Cases". Use cases are frequently asked for how-to scenarios and workarounds. The main README is basically a Table of Contents with links. I have also added 2 use cases, Portfolio Management and Resource Management.

This is is not associated with a specific work package, but rather a value-add to customers