Allow "User" custom fields to act as filters in global work_packages
Custom fields of type "User" and "Version" were excluded in 2018 from
appearing in the global work_packages reports. I guess that's because
their values are set per-project. However, it seems that the current
"principal" endpoint returns the global list of users now, so that a
filter of type "User" works correctly, at least in my installation.