avoid left joining on roles to avoid multiple projects returned
As the permission check used to be left joined directly to the list of projects, with a user having multiple roles, this lead to having a project returned multiple times.

The permission check is not necessary at this point at all however since the query will already add a visible scope if necessary.