fix flickering test
because of ng-select/ng-select#1869 we can't use a string with spaces in ng-select input value while the list of items is empty. The spaces would be ignored and matching an entry would then fail. Instead wait for ng-dropdown-panel to appear which means items are present, the bug is inactive, and the spaces can be properly sent to the input.

flickering test was spec/features/custom_fields/multi_user_custom_field_spec.rb:61