[41437] Use signalling on project edit field (#10432)
* Add project-edit-field that uses op-autocompleter

* Replace for_type param with regular filter

* Fix unrelated spec now fixed in behavior due to the select helper change

The spec did not apply the new type previously, so would not override
the description. This caused the expects to remain valid, but was
actually incorrect behavior.

* Ensure we're not focussing the project field on inline create

* fix flickering test

because of ng-select/ng-select#1869 we can't use a string with spaces in ng-select input value while the list of items is empty. The spaces would be ignored and matching an entry would then fail. Instead wait for ng-dropdown-panel to appear which means items are present, the bug is inactive, and the spaces can be properly sent to the input.

flickering test was spec/features/custom_fields/multi_user_custom_field_spec.rb:61

* fix tests broken in 65695f1

- Reintroduce sleep to make test pass. I'd like to get rid of it but
  can't understand why it's needed.
- refine results_selector only if it is 'body'.

* fix tests by not waiting for dropdown before typing for global search

Co-authored-by: Benjamin Bädorf <b.baedorf@openproject.com>
Co-authored-by: Christophe Bliard <c.bliard@openproject.com>