clear Rails cache between tests to avoid Settings leaking state
Settings are cached with a key having the timestamp of the latest
updated setting. When no settings are saved, current time is used
instead. For fast tests occurring in the same second, the settings state
is kept as they share the same cache key. That leads to non-isolated
tests and possibly failures. Clearing the Rails cache before each test
avoids it.