Change alignment for editable toolbar titles
This is a rather dirty but workable global change to the CSS that moves editable toolbar titles to be on the same
vertical alignment when compared to the rest of the page. However, when editing or hovering these titles, a box appears
around them (indicating an input field). Since the previous CSS assumed paddings farther up the HTML tree and the
scrolling to happen inside these paddings, we've had to move the paddings further down the HTML tree, slightly
complicating how they're applied. However, this makes the change possible with minimal changes.