[41718] Filter query filter values in the frontend
If we loaded all values in the first run, filter them in the frontend.
The backend does not provide queries and typeahead filters for all
resources, and we can avoid additional requests if we just filter them
here.

Note that if more than the max API size items are returned, and the
backend does not provide any filtering, this will result in all items
being shown at all times.

https://community.openproject.org/wp/41718