Also request notifications from inactive tabs at much larger intervals
We have seen reports from customers where requests of notifications fail
whenever they re-activate the window after leaving it for some time.

This is probably due to some proxy or network timeout in between, but
notifications is the first request they see and connect to them.

This PR allows notification requests to be performed every 2 minutes
while the tab is inactive. This has the side-effect of a ping request,
keeping the session alive as well as preventing a 10s delay between
reactivating the tab and seeing new notifications come in.