deserialize empty string setting value to nil when not string format
the seeder blindly saves all settings to database, setting the value to empty string when no value is defined. When the setting is read back, it can be misinterpreted. For instance a boolean setting with a nil value will have the value '' once seeded, and it is truthy which is wrong.

add a helper to restore settings after resetting them in tests