[41932] Use filter ID instead of _type which are always available
When creating a new query filter instance from the frontend, the `_type`
is not set. The same is true for when filters are parsed from a form
request.

However, for identifying which filter is being used, we can simply use
the `id` property which contains a unique filter ID such as `assignee`
and `project`.

This prevents flickering in the team planner while we already have an
assignee filter, but one that does not have a _type yet.

https://community.openproject.org/wp/41932