[41718] Allow ng-select internal filter when no backend API involved
When ng-select gets a [typeahead] input, it rightfully assumes we are
filtering on the backend. However, in our case we always input a
typeahead and just replay the items again.

I think this is a flaw of op-autocompleter, but it's not easily fixable.
For 12.1 I suggest to avoid passing in a typeahead in this case and
refectoring op-autocompleter afterwards (already discussed this with
Ben).

https://community.openproject.org/wp/41718