Account for padding in scroll height
The projects dropdown is lazily rendered. A new page is added whenever
the scrollbar reaches the bottom of the scrollable area.

A padding of 6px was introduced, which results in the scrollable area never reaching the true bottom of the page.
This in turn results in only the first 250 pages being rendered.

https://community.openproject.org/work_packages/42058