Fix project include list loading flicker
When opening the project include list the loading indicator would be removed before the list of projects would be
rendered. This turned out to be due to the interaction of several different observables firing at different times,
causing the loading indicator to be removed before the project list was filled. This PR fixes that, making sure the
loadign indicator will not get removed before the project list is actually filled

Co-authored-by: Oliver Günther <o.guenther@openproject.com>