Load projects via signalling in project-autocompleter
This changes the projects loading logic aroundto use signalling. After
loading the list of projects, the data gets transformed into the same
tree structure that's in use in project-include.

Afterwards, we flatten the tree back to a list so that ng-select can
consume it. The hierarchy is "faked" with paddings.