Consolidate two feature specs into one
OAuth client settings cannot well tested without testing
storages. That part of the administration has some
characteristics of a dialogue. Many labels of buttons
change, for instance, if the storage already has
OAuth client details added or not. We want these tiny
changes in the UX to be covered by this spec.