Fix mailer interceptors rejecting empty to addresses
In case of invalid mails or trying to send mails to builtin users
like the system user, there is an interceptor in place trying to reject
sending the mail as the to address is empty.

This interceptor however was no longer loaded, as the ActionMailer configuration
is loaded before it gets added.

Instead, explictly add the interceptors

https://community.openproject.org/work_packages/42377