Fix spacing of "Log time" button in the my spent time widget
Changes to the way full calendar CSS was being overwritten caused the
date display to overlap the log time button on the right hand side.

This commit changes the behavior entirely by putting the log time button
above the full calendar toolbar on mobile. On desktop, some spacings
are improved so that the date feels more centered.

Closes https://community.openproject.org/work_packages/42085/activity