Add cop to enforce usage of ServiceResult factory methods
The cop is named `OpenProject/UseServiceResultFactoryMethods`. It will
check for ServiceResult.new call and check if it can be replaced by
`ServiceResult.success` or `ServiceResult.failure`. These factory
methods make the code easier to understand by being shorter and more
explicit.

This is the first cop I write, so there is probably room for improvement.