reduce count calls in collection representer
will_paginated, when actually fetching the records will do an SQL count on top of the fetching. We don`t need that
in our offset paginated represters since those implement their own counts.

And if the per_page is 0, we don`t even have to try to fetch any objects so we can use `none` to avoid the database
roundtrip.