quickfix for principal filter performance
Instead of instantiating the users only to then get the name and the id for the available values of principal based filters,
only pluck the id directly as that is the only part of the allowed_values that is used nowadays.

Sorting is not necessary at all any more since the allowed_values are no longer displayed within the application.