Refactor user-autocompleter
The old user-autocompleter was incompatible with angular reactive forms,
did not offer the ability to hand over a full json object as the initial
value (which caused the name to be missing when rendering from the
backend), and was not conforming to our naming standards because it
missed the `op-` prefix. These things are all fixed in this commit.

Co-authored-by: Dmitrii Batracov <dbatracov@n-ix.com>