Fix/40286 fixes 1 (#10872)
* Reduce spacing per level in op-project-autocompleter

This used to be 20px per level, but is now reduced to 16px, or 1rem, to
align with `spot-list`

* Fix op-project-autocompleter appendTo and disabled reason

The disabled reason text was shown even if the item was not disabled.
This is fixed here. Also, the op-project-autocompleter appended the
results modal to `body` by default, but that screws up scrolling
oftentimes, so here it is changed back to open in-context by default.

* Disable parent projects of search results

In project-include, parent projects of search results were still
clickable. This behaviour was not to spec, so we now disable projects
that were not directly search results.

* Disable projects that are not search results

In the global project select drop-modal,
all projects were clickable before this.
However, when a project was shown only because
child projects matched the search criteria, it
should be disabled. This commit fixes that.

* Always open project select on top in notification settings

* Fix styling issues in op-project-autocompleter

* Fix text color for list items in project drop modals

* Auto-open project dropdown in work package table

* Fix tooltip being cut off by overflow: auto in project list

* Fix failing spec

* Fix failing spec

* Fix linting errors

* Fix onchange for empty project select on work package move page

* Fix linting errors

* Fix the way focus works in inline create

* Adapt spec to match disabled item

Co-authored-by: Oliver Günther <mail@oliverguenther.de>