Validate storage host to not redirect to "/index.php"
If the storage is a Nextcloud instance and it redirects
to a new path containing "/index.php" then it is a strong
indicator that the Nextcloud server was not fully set up.
It usually means that the Apache is not fully configured
and thus the Bearer authorization tokens get stripped
from API calls.

https://community.openproject.org/work_packages/43297