Work around for project dropdown overflows
The new hierarchical project dropdowns caused an issue because long
project names are cut off by default. This makes it impossible to
distinguish when the naming conventions puts the difference between
project at the end of their name, for example:

"Project with long name 1" and "Project with long name 2" will both
show up as "Project with...".

This commit adds a `title` attribute with the name, so hovering the
option will give users the possibility of reading the full name.

It also widens the result list in a rather hacky way. It works, but
could cause the result list to be cut off by the screen edge if
the project select is e.g. all the way on the right in the wp table.

Closes https://community.openproject.org/projects/openproject/work_packages/43083/activity