Fix internal server error when scheduling a wp
When rescheduling a work package, in very specific cases the code would
hit the `set_schedule_service`, with a work package without a duration.
The service would attempt to set new wp dates by calculating the end date
from provided start date. However, if the work package has no duration
this calculation fails.

This commit defaults to using a 0 day duration if a duration is not set.

Closes https://community.openproject.org/projects/openproject/work_packages/40921/activity