Fix project name overflow in app header (#11028)
* Fix project name overflow in app header

Longer project names were overflowing in the app header, causing the
text to flow behind the logo, search and user menu. This commit fixes
that issue. This also fixes the menu toggle shrinking in width whenever
there was too little space. Instead, the project title is shrunk.

Closes https://community.openproject.org/projects/openproject/work_packages/43314/activity

* Give header project selector all available space

On mobile and tablet the instance logo dissapears from the app header,
but the left and right side were both still taking 50% of available
space. However, it makes sense that the project selector can use this
screen estate to show longer project names.