[#43530] Apply "No SSL" as specified with packager
Default value of `Setting.rails_force_ssl` was `false` before
9e3c51299ae, and `true` after. As the default value changed to true, it
lead a packaged installation to a forced ssl even if "No SSL" was
chosen.

Instead of unsetting `OPENPROJECT_RAILS__FORCE__SSL` and assume a
default value of `false`, explicitly set it to either `true` or `false`
to fix the problem.