Fix flickering datepicker feature test
failing test was ./spec/features/work_packages/details/date_editor_spec.rb:134

When setting date in tests through `Components::Datepicker#set_date`,
the month viewed was changed by clicking on next month two times, and
then the day is clicked. It can occur that the debounce on
`dateChangedManually` triggers right between choosing another month and
clicking the day, thus resetting the date picker to display the initial
month of the date.

When that happens, the day element found on the month displayed before
the date picker gets refreshed is now stalled and can not be clicked
anymore.

The proposed fix is to set the debounce delay to 0 initially, so that
everything triggers immediately in `ngAfterViewInit`, and then set the
debounce delay to its real value for subsequent calls.