Also scroll to first unread activity on first mobile route
If we're opening a work package link on the notification email, we're
routed to /notifications/details/ID/activity.

On that route, https://github.com/opf/openproject/pull/11020 prevented
scrolling to the first unread activity unless we're coming from
notifications.

For a deep link however on mobile, the scroll behavior should also be
retained.

https://community.openproject.org/wp/43311