Do not derive a field set to nil in request
With a work package having only start_date field set (other are nil)
If duration is set to 3 and due_date is set to nil
Then two fields are set, and one in unset, which is not allowed.
As both duration and due_date have been explicitly set, start_date is
the one which must change. It must be set to nil.