implementing `Journal.at_timestamp`
in order to retrieve journal entries that correspond to a certain point in time.

If you have journals for Monday, Wednesday, and Friday, then `Journal.at_timestamp(Thursday)` will only include the Wednesday journal because that one is the most current on Thursday.

This is a step towards the baseline-queries feature:
- https://community.openproject.org/projects/openproject/work_packages/43483
- https://community.openproject.org/projects/openproject/work_packages/26448