Fix opAutofocus directive
The opAutofocus directive was not setting focus if used like a normal
`autofocus` HTML attribute:

```
<input opAutofocus />
```

The above example would not produce a focus, because counterintuitively
Angular treats the default input to the directive here as an empty
string instead of undefined. This causes the default value of `true` NOT
to be set. Instead, the empty string was treated like the falsy value
it is.