Pass ssl_verification value to the IMAP library
The current implementation sets the `verify_mode` based on the `ssl_verification` value and doesn't pass the value to the library "net-imap", which sets the variable `verify` to the default value `true`.
However, in net-imap library the `verify_mode` is set/reset inside the `create_ssl_params` function based on the value of `verify` which is set to true as a default value.

My fix is not to set the value for `verify_mode` and instead pass `ssl_verification` to the net-imap library which handles the settings correctly.

Signed-off-by: Oliver Günther <o.guenther@openproject.com>