Do not move followers backwards when rescheduling
See https://community.openproject.org/wp/44053

When moving work packages, we try not to move followers except if forced
to satisfy follows relations. In this case the followers are moved
forward.

The work packages in precedes/follows relations can never be moved to an
earlier date automatically. Only the user can move backwards.