Chore/modal spacing (#11314)
* [chore] reworked spot-container spacings

* [chore] applied spot-container to spot-modal--body

- added style override for spacing within modal body

* [chore] removed unused modal footer class

* Remove scrollbar from project-include

Project include was creating scrollbars for the project list even if
it was not needed. This was caused by tooltips that were hidden in the
background but still affecting layouting.

This commit hides the tooltips correctly, and aligns the HTML structure
of project include to match that of file link list. A minimum height is
added to the project include list to make sure tooltips are readable even
if there is only one item in the list.

* [chore] rework of datepicker modal

- changed spacings, removed inner margins
- changed anchors to buttons (spot-link)
- added flatpickr override
- removed min-height
- moved project include min-height to root element

* [chore] fixed test selector

Co-authored-by: Benjamin Bädorf <b.baedorf@openproject.com>