Import HttpClient module only once (#11370)
* Import HttpClient module only once

The reimport caused our http interceptor to not be set for the second
httpclient module. That in turn caused missing CSRF headers on at least
one call, which resulted in a 422 and a subsequent refresh of the session
and the accompanying logout.

Closes https://community.openproject.org/work_packages/44161/activity

Co-authored-by: Oliver Günther <mail@oliverguenther.de>

* Import HttpClient module only once

The reimport caused our http interceptor to not be set for the second
httpclient module. That in turn caused missing CSRF headers on at least
one call, which resulted in a 422 and a subsequent refresh of the session
and the accompanying logout.

Closes https://community.openproject.org/work_packages/44161/activity

Co-authored-by: Oliver Günther <mail@oliverguenther.de>

* Add HttpClient mock to specs

Co-authored-by: Oliver Günther <mail@oliverguenther.de>