improve visiblity scope performance by EXIST
IN in combination with a subquery can lead to very poor performance if that subquery ends up being executed once for every candidate record. EXIST typically performs better and only rarely worse. Using a join would be even stable but this is harder to do in the visible scopes which end up being used in a lot of queries combined of various scopes.