`Query::Results`: search work packages at all given timestamps
The result includes work packages that match the filter criteria at *any* of the given timestamps. In the context of the baseline-comparison feature, this means that the result contains the work packages that do match the filter criteria today, as well as the work packages that have matched the filter criteria at the baseline date.