Fix font loading in storybook
Both Lato and the icon font were not loading correctly since the CSS
expects them to be available under a certain URL. This URL is available
in the core app, but not storybook. This commit makes sure the assets
load correctly in Storybook, giving us pretty icons and the right font.