Fix viewpoint seeding in multitenancy
As soon as the first attachment is saved, a notification delayed job is now
triggered.

That in itself is not a a problem, however the CreateTenantJob calls
`DelayedJob.delay_jobs = false` which results in them running inline.

At the same time, an `Apartment::Tenant.switch` happens within the job
to get to the correct schema. This switch includes a
`RequestStore.clear!` which removes the current_user value.

That is why in multitenancy, seeding the first viewpoint works, then the
next viewpoint is built with the AnonymousUser, resulting in failing
permissions.