Fix custom field opening
Custom fields with multiselection that were added to the end of the work
package form were dissapearing completely when opening. This was caused
by code attempting to use `evt.originalEvent`. This attribute is part of
the jQuery Event which wraps the native browser event. However, the
event being passed in was a native browser event.

Closes https://community.openproject.org/work_packages/44846/activity