Fix positioning for preview modal
After the spot-modal plumbing rewrite, the preview modal was not being
positioned correctly when opened a second time. It was a combination
of two things; when opening the second time, the contents of the modal
are already cached, and thus the reposition call that results from
updated content was faster than the reposition call from the preview
trigger service.

The actual bug was in that specific reposition call; it was using the
wrong elements to do the positioning; namely one level down from where
it should be happening.

Closes https://community.openproject.org/work_packages/45058/activity