Remove pry-stack_explorer gem
It conflicts with pry-byebug, making up/down commands not working after
stepping into a method.

Also pry-stack_explorer makes the first 'up' command enter the pry code
frames, which is 'down' the current frame, so it is surprising. One has
to enter up multiple times before coming back to the initial frame. It
works as expected when only pry-byebug is used.