Fix flaky test
When the team planner gets rerendered the drop down will be closed. If it happens at the bad time, the test will keep trying to find an element that has been closed.

Add the click in the retry block so that we try to open the dropdown everytime (clicking twice does not close it. It keeps it open)