[#3897] Fix error message
Someone having the role called "Project admin" may not have the
permission to archive subprojects. Having an error message asking to
contact the project administrator will leave them puzzled as they are
the project administrator.

The real person able to help here is the project administrator. Reflect
this in the error message.