cost query - filter by cf value but return id
The filtering has to happen by custom option value since that is what gets sent from the frontend and there are also persisted queries having the value. But the id of the custom option needs to be returned by the cost query statement as that is used for looking up the casted value later on. Just for list cfs, it would be possible to avoid this indirection but since the rest of the values also need to be casted, all values need to be treated the same.