at_timestamp: correct mapping of the `updated_at` column
The `work_package_journals` table does not have an `updated_at` column as the `work_packages` table. When asking for the `updated_at` attribute, we need to map this to the `journals` table.

Originally, I have mapped the `updated_at` attribute to `journals.created_at`, assuming that each journal entry corresponds to a specific change. However, the journal entries can be updated when several changes are made in short succession in order to keep the activity log cleaner. Because journal entries are updated in this scenario, the last time of update is actually `journals.updated_at` rather than `journals.created_at`.

https://github.com/opf/openproject/pull/11243#issuecomment-1386919337