signaling for notifiction api
This was intended to improve the performance of the most often used
request, the one to query for the existence of new notifications.
However, there is no performance gain to be measurable so that idea was
abandoned.

The reason for this is that we have already an optimized case for the
default representer when a pageSize of 0 is provided.