Hide next and prev months days from single datepicker
They are already hidden by default in ranged datepicker, but for
showMonths=1, flatpickr behaves differently and adds a lot more days

To circumvent that, hide the previous month days but keep visible.
For next month days, this won't work as they result in more whitespace
added.

That's why we have to override the styling of flatpickr to use a grid
with better positioning of the days, to prevent shifts to layout when
some of the columns are missing

https://community.openproject.org/WP/46189