Fix alignment of text in project tables
Text in table cells was spaced too high up. Icons generally looked
correct. Changing the paddings for the cells was out of the question,
because that would misalign items that fill the full height.

This commit pushes down `a` and `span` elements inside table cells by
one pixel. This is rather hacky, but it makes sure we don't have to
completely redo the way positioning works in table cells in general.

Closes https://community.openproject.org/work_packages/45946/activity