Fix highlightedAttributes null error on query errors
The UrlParamsHelperService.encodeHighlightedAttributes method was incorrectly mapping the highlightedAttributes
id into the url parameters. The id is not present when the HalResource is not loaded. We should use the href
instead in order to compose the id, href is present in the HalResource even when the resource is not loaded.