add non null constraints on data_id/type polymorphic foreign key
An OP instance reported a lot of records where both are null. It is still unclear how this has happend.
Having an explicit not null is not as good has having a proper foreign key constraint, which however cannot be modelled on polymorphic
relations in SQL. And the invalid records reported by the instance should at least be prevented.