Fix word breaking on WP pages
This commit touches the way words are broken in wp subjects, descriptions,
custom fields, and activity messages.

Previously, words would get broken randomly at the end of a line. This
behaviour is now improved. Words tend not to get broken unless there is
no space at all for the word to fit on a line without being broken.

However, this commit does not enable hyphenation for these places,
because `hyphens: auto;` causes the browser to break words more
aggressively.