Make some feature specs faster
`raise_if_found` helpers were uselessly waiting 5 seconds for the
element to show up. Using regular capybara matchers saves the waiting
time.

Also, `close_autocompleter` was waiting for it to be visible, or even
opening it to be able to close it. If it is not open, returning early
saves time too.

About `ensure_value_is_input_correctly`, it is more optmistic: it first
tries without waiting, and if the input value is not correctly set, it
waits as it did before.

For `spec/features/work_packages/table/queries/filter_spec.rb`:
- before: 2 minutes 41 seconds
- after: 1 minute 15.81 seconds