Improve performance of querying with timestamps
The performance improvement is on `/api/v3/projects/:id/work_packages`
endpoint with `timestamps` parameter.

The inner queries (one per timestamp) are filtering on given project id
and subprojects ids, but the projects table was only available in the
outer query, leading to bad performance.

The perf got better once the projects table is joined in the inner
queries. In fact, the projects table should have been automatically
joined by ActiveRecord as the query is calling `includes(:project)` but
as `Journable::HistoricActiveRecordRelation#eager_loading?` is
overridden to return `false`, joining the eager loaded tables was not
done. That's why it is being done explicitly.

The performance improvement is multiple order of magnitude: on a test
database, it went from more than 2 minutes to 200 ms. The overall api
endpoint query takes less than 2 seconds to complete.