Remove ability to add groups to projects for seeds
It was introduced in PR #7495.

Then groups were removed from the seeding files in c8fda72e17. Only the
example file still had `groups`.

Then it was reintroduced again in 3a44b38 but only the `name` was used.
The `projects` property has not been used for 4 years and is dead code
that awaited to be pruned.