check visibility at timestamps
The visiblity checks for timestamp based filtering is moved down into
the `Journable::WithHistoricAttributes` delegator. This enables
scenarios where a work package used to be visible but no longer is and
vice versa.

If a work package is no longer visible, but part of the query results,
permission is rechecked for every timestamp provided. If a work package
is visible now, no visibility check is conducted for the individual
timestamps.

On the representer layer, if a now invisible work package is to be
rendered, it is treated as if it were deleted. Only the bare minimum of
information is returned. On the other hand, the `attributesByTimestamp`
will contain all the attributes, regardless of whether they changed or
not. This is done so that the former attributes can be displayed without
revealing the current attributes.