Make xml to hash conversion faster by using Ox
The hash produced by Ox is very different from the one produced by
`Hash.from_xml`. A big part of the new code is about converting from Ox
hash format to ReXML hash format so that the rest of the code works
seamlessly. The conversion code is utterly complex. Sorry for that.