refactor: Extract spec uploading code in a class
The flaky test fixed 59eb308cbb was failing because the test was not
waiting for the attachments element to be displayed. In some other
similar tests, the code was waiting for it, and one test ever scrolled
to that element.

If code was properly factored, the above test would never have been
flaky.

This commit is an attempt to factor the interactions with the
attachments list so that the next test written using it will be reliable
by default.