Disable already selected projects in autocompleter (#12402)
* Disable already selected projects in autocompleter

Before this fix, projects that were already selected could be selected
again in the `op-project-autocompleter` component. By default, ng-select
will remove items from the provided list if it finds they have already
been selected, but this does not work for our usecase, as it would break
the tree view. Instead, this commit makes it so that the list of
projects gets updated whenever the value changes, disabling selected
projects.

Closes https://community.openproject.org/projects/openproject/work_packages/47074/activity?query_id=3922