Remove dead code in seeding
The ability to add a link to a child work package in a description with
the `##child:n` syntax was added and used in 630572fba33aa in 2018.

When there was new seed data for "demo-project" in PR #9714 in 202, the
usage of `##child:n` was removed (commit 7c87bfeab7) and never used
again since.