Refactor ProjectRepresenter to stop referencing Projects::Status attributes
As a step in the direction of refactoring onto the new columns in the
projects and project_journals tables, we modify the attributes
referenced by project_representer to those directly in the projects table.

**NOTE**: Spec for asserting the cache invalidation when Projects::Status
is changed was removed since this is no longer an association that will
be tied to projects.