Add alias conversion from `status_code` to `status`
Following in line with what WorkPackages has for
`done_ratio` -> `percentageDone`, this commit adds the conversion
capability for mapping `status_code` out to `status`.

**NOTE**: Curious of the implications of adding this to the conversions
constant. I'd like to think there aren't any ramifications beyond this
given that in essence, we'd always refer to inbound `status` keys as
`status_code` with the ongoing work.