Skip tests whose needed CLI command is unavailable
Mark a test to be skipped with `skip_if_command_unavailable: 'some_command'`
or `skip_if_commands_unavailable: ['some_command', 'another_cmd']` in a
group or example metadata.

Tests won't be skipped in CI environment, and will fail as expected if
the needed CLI command is not available.

For instance, the `spec/models/ldap_auth_source_spec.rb` spec starts a
LDAP server before running. It relies on `java` to start the ldap
server. If the `java` command is not available, the tests marked with
`skip_if_command_unavailable: 'java'` will be skipped.