Fix how local time zone is retrieved
On CI, the previous way of getting the local time zone was returning
'Edinburgh' because local time is UTC, the UTC offset is 0, and
`ActiveSupport::TimeZone[0]` returns the Edinburgh time zone. But
Edinburgh is not in our list of assignable time zones and so it would be
ignored by `Notifications::ScheduleDateAlertsNotificationsJob`, making
the test fail.

Getting the first assignable time zone with an identical UTC offset is
more reliable.