Maintain constructor signature (*classes)
In order to keep this constructor flexible and be able to simply
provide class names that warrant replacing, performed a refactor
on the internal works.

I don't believe recursion is strictly necessary and if anything, I feel
a simple "hijack" or "focus" on replacements one by one without the need
for instantiation is simpler to wrap my head around.

Made sure to add some specs to cover both constructor signatures and
expected behavior.