Rename "Professional Support" links to "Enterprise Support"
The link to which a user is redirected when clicking on Enterprise Support
links is determined by whether or not they are on the Community or Enterprise Edition.

A user in Community Version gets redirected to the pricing page, while a user
on the Enterprise Edition gets redirected to the new landing page for Enterprise
Support docs.