Restore some authentication settings in admin
Also makes ckEditorAugmentedTextarea read-only if the wrapped text area
is disabled.

As for the invitation_expiration_days setting, commit 269f9416ddc and
commit 286d862c38 got intertwined with merge commit eb63698a3dd, and
resulted in hiding the setting when
`OpenProject::Configuration.registration_footer` is set.

As for the registration_footer setting, PR #6321 intent was to prevent
modifying it if it was already set via env var or configuration file.
When `OpenProject::Configuration` and `Settings` were merged, this code
was not updated. With the default value being non-blank, it lead to
hiding the setting from the administration pages.

This commit restores both settings.