Fix deserialization issue with Settings
It was preventing deployments as Setting.rails_cache_store was set via
an env variable but failed to validate as the allowed values were
symbols and the value was a string. As it was not converted to symbol
before the validity check, it was not valid and the application could
not start.