Fix test
When starting with an empty array, the search results options would be
displayed before getting the results, leading to a search result option
to have the focus when pressing enter (only for a brief moment, but long
enough to see the focus flicker).

Removing the empty array elements fixes it.